fix: reactor zuri training routes
