--- routes/zuriTrainingRoute.js
@@ -1,11 +1,0 @@
-const express = require('express');
-const {
-  createIntern,
-  internApplicationValidationRules
-} = require('../controller/zuriTrainingApplicationController');
-
-const router = express.Router();
-// Admin routes
-router.post('/zuri/training/application', internApplicationValidationRules(), createIntern);
-
-module.exports = router;
